Migrated imdb plugins into a component
