+++ flexget/api/__init__.py
@@ -4,5 +4,5 @@ from .app import api_app, api, APIResource, APIClient  # noqa
 from .core import authentication, cached, database, plugins, server, tasks, user, format_checker  # noqa
 from .plugins import (
     tvdb_lookup, tvmaze_lookup, tmdb_lookup, status, pending,
-    imdb_lookup, rejected, schedule, variables  # noqa
+    rejected, schedule, variables  # noqa
 )
