+++ flexget/plugins/input/rlslog.py
@@ -1,19 +1,27 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import logging
 import re
 import time
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 from requests import RequestException
 
 from flexget import plugin
 from flexget.entry import Entry
 from flexget.event import event
-from flexget.utils.imdb import extract_id
+from flexget.utils.cached_input import cached
 from flexget.utils.log import log_once
 from flexget.utils.soup import get_soup
-from flexget.utils.cached_input import cached
+
+try:
+    # NOTE: Importing other plugins is discouraged!
+    from flexget.components.imdb.utils import extract_id
+except ImportError:
+    raise plugin.DependencyError(
+        issued_by=__name__, missing='imdb',
+    )
+
 
 log = logging.getLogger('rlslog')
