+++ flexget/plugins/metainfo/nfo_lookup.py
@@ -1,12 +1,20 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import logging
 import os
 import xml.etree.ElementTree as ET
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 from flexget import plugin
-from flexget.utils.imdb import is_valid_imdb_title_id
+
+try:
+    # NOTE: Importing other plugins is discouraged!
+    from flexget.components.imdb.utils import is_valid_imdb_title_id
+except ImportError:
+    raise plugin.DependencyError(
+        issued_by=__name__, missing='imdb',
+    )
+
 from flexget.event import event
 
 log = logging.getLogger('nfo_lookup')
